feat:add solution to esm.mjs
